requirement.txt
playwright
python-dotenv
pandas
